Move form elements to form directory
