+++ src/resources/form/abstract/collapsiblefield.js
