+++ src/resources/form/optionfield-checkbox.html
